make a navbar Component
